Changed timers to use ScheduledExecutorService instead
