+++ src/model/Model.java
@@ -554,6 +554,7 @@ public class Model {
 	}
 	
 	public void refresh(){
+		System.out.println("Refreshing...");
 		ServiceManager.getInstance().restartModelUpdate();
 		applyChangeMethods();
 	}
