+++ src/service/updateservice/ModelUpdater.java
@@ -2,8 +2,6 @@ package service.updateservice;
 
 import java.util.Date;
 import java.util.List;
-import java.util.Timer;
-import java.util.TimerTask;
 
 import org.eclipse.egit.github.core.IRepositoryIdProvider;
 import org.eclipse.egit.github.core.Issue;
@@ -12,9 +10,6 @@ import org.eclipse.egit.github.core.Milestone;
 import org.eclipse.egit.github.core.User;
 
 import service.GitHubClientExtended;
-import ui.UIReference;
-import ui.components.StatusBar;
-import util.events.RefreshDoneEvent;
 import model.Model;
 
 public class ModelUpdater {
@@ -23,12 +18,6 @@ public class ModelUpdater {
 	private CollaboratorUpdateService collaboratorUpdateService;
 	private LabelUpdateService labelUpdateService;
 	private MilestoneUpdateService milestoneUpdateService;
-	private long pollInterval = 60000; //time between polls in ms
-	private Timer pollTimer;
-	private int stopwatchInterval = 1000;
-	private int timeRemaining = 60;
-	private static final int SECS = 60;
-	private Timer stopwatch;
 	private Date lastUpdateTime = new Date();
 	
 	public ModelUpdater(GitHubClientExtended client, Model model, String issuesETag, String collabsETag, String labelsETag, String milestonesETag, String issueCheckTime){
@@ -43,7 +32,7 @@ public class ModelUpdater {
 		return lastUpdateTime;
 	}
 	
-	private void updateModel(IRepositoryIdProvider repoId){
+	public void updateModel(IRepositoryIdProvider repoId){
 	    updateModelCollaborators(repoId);
 	   	updateModelLabels(repoId);
 	  	updateModelMilestones(repoId);
@@ -97,50 +86,6 @@ public class ModelUpdater {
 			}
 		}
 	}
-	
-	public void startModelUpdate(){
-		if(pollTimer != null){
-			stopModelUpdate();
-		}
-		pollTimer = new Timer();
-		
-		// get the current repo id from the model now so that the updates done will correspond with the current id in case of project switching
-		final IRepositoryIdProvider repoId = model.getRepoId();
-		TimerTask pollTask = new TimerTask(){
-			@Override
-			public void run() {
-				updateModel(repoId);
-				UIReference.getInstance().getUI().triggerEvent(new RefreshDoneEvent());
-			}
-		};
-		pollTimer.scheduleAtFixedRate(pollTask, 0, pollInterval);
-		
-		stopwatch = new Timer();
-		TimerTask countdown = new TimerTask() {
-			@Override
-			public void run() {
-				StatusBar.displayMessage("Next refresh in " + getTime());
-			}
-		};
-		stopwatch.scheduleAtFixedRate(countdown, 0, stopwatchInterval);
-	}
-	
-	private int getTime() {
-	    if (timeRemaining == 1) {
-	        timeRemaining = SECS;
-	    } else {
-	    	--timeRemaining;
-	    }
-	    return timeRemaining;
-	}
-	
-	public void stopModelUpdate(){
-		if(pollTimer != null){
-			pollTimer.cancel();
-			pollTimer = null;
-			stopwatch.cancel();
-			timeRemaining = SECS;
-			stopwatch = null;
-		}
-	}
+
+
 }
